os.c: use size_t for string size

+++ src/os.c
@@ -11,7 +11,7 @@
 void
 OS_fix_filename ( char * dirty )
 {
-	for ( unsigned i = 0; dirty[i] != '\0'; ++i )
+	for ( size_t i = 0; dirty[i] != '\0'; ++i )
 		if ( ( ( dirty[i] & 0xC0 ) != 0x80 ) && ( dirty[i] == '/' || dirty[i] == '\\' ) )
 			dirty[i] = '_';
 }
